ScenarioUser controller: fix tests, add validation, add stricter distinction between id and user_id

+++ app/controllers/api/v3/scenario_users_controller.rb
@@ -3,11 +3,10 @@ module Api
     class ScenarioUsersController < BaseController
       respond_to :json
 
+      # Only scenario owners, having the delete authority, may manage scenario users
       before_action { doorkeeper_authorize!(:'scenarios:delete') }
 
       before_action :find_and_authorize_scenario
-
-      before_action :validate_users_presence, only: %i[create update destroy]
 
       def index
         render json: users_return_values
@@ -27,9 +26,9 @@ module Api
               status: :unprocessable_entity and return
           end
 
-          # Send an invitation if stated in the params
+          # Send an email invitation to the added user only if stated in the params
           if permitted_params.dig(:invitation_args, :invite) == true
-            send_invitation_mail(scenario_user)
+            send_invitation_mail_for(scenario_user)
           end
         end
 
@@ -41,30 +40,23 @@ module Api
       end
 
       def update
-        scenario_user_error = nil
-        http_error_status = :unprocessable_entity
+        return unless validate_scenario_user_params(['role', ['id', 'user_id', 'user_email']])
+
+        http_error_status, scenario_user_error = nil, nil
 
         permitted_params[:scenario_users].each do |user_params|
-          if user_params.try(:[], :role).blank?
-            scenario_user_error = user_params.merge({ error: 'No role given to update with.' })
-
-            break
-          end
-
-          scenario_user = find_scenario_user_by_params(user_params)
-          if scenario_user.blank?
+          # Find the user
+          unless (scenario_user = find_scenario_user_by_params(user_params))
             scenario_user_error = user_params.merge({ error: 'Scenario user not found' })
             http_error_status = :not_found
 
             break
           end
 
-          unless (
-            scenario_users_updated = scenario_user.update(
-              role_id: User::ROLES.key(user_params.try(:[], :role).try(:to_sym))
-            )
-          )
+          # Attempt to update the user
+          unless scenario_user.update(role_id: User::ROLES.key(user_params[:role].to_sym))
             scenario_user_error = scenario_user.errors
+            http_error_status = :unprocessable_entity
 
             break
           end
@@ -78,36 +70,17 @@ module Api
       end
 
       def destroy
-        param_user_ids = permitted_params[:scenario_users].pluck(:id).compact.uniq
-        param_user_emails = permitted_params[:scenario_users].pluck(:email).compact.uniq
-        user_ids = User.where(email: param_user_emails).pluck(:id)
+        return unless validate_scenario_user_params([['id', 'user_id', 'user_email']])
 
+        # Find scenario_users by id, user_id, user_email,
+        # or the id of users found through their email address indirectly
         scenario_users = @scenario.scenario_users.where(
-          'user_id IN (?) OR user_id IN (?) OR user_email IN (?)',
-          user_ids, param_user_ids, param_user_emails
+          'id in (?) OR user_email IN (?) OR user_id IN (?) OR user_id IN (?)',
+          permitted_params[:scenario_users].pluck('id'),
+          permitted_params[:scenario_users].pluck('user_email'),
+          permitted_params[:scenario_users].pluck('user_id'),
+          User.where(email: permitted_params[:scenario_users].pluck('user_email')).pluck(:id)
         )
-
-        # If we found less users than requested, we either:
-        # a) Could not find one of the requested users, or
-        # b) There are duplicate entries in the request that we filtered out
-        if scenario_users.count < permitted_params[:scenario_users].length
-          missing_ids = param_user_ids - scenario_users.pluck(:id)
-          missing_emails = param_user_emails - scenario_users.map { |su| su.user.email }
-
-          if missing_ids.present? || missing_emails.present?
-            render \
-              json: {
-                error: "Could not find user(s) with id or email: #{(missing_ids + missing_emails).join(',')}"
-              },
-              status: :not_found
-          else
-            render \
-              json: { error: "Duplicate user ids found in request, please revise." },
-              status: :unprocessable_entity
-          end
-
-          return
-        end
 
         scenario_users.destroy_all
 
@@ -117,7 +90,11 @@ module Api
       private
 
       def permitted_params
-        params.permit(:scenario_id, scenario_users: [%i[id role email invite]], invitation_args: %i[invite user_name id title])
+        params.permit(
+          :scenario_id,
+          scenario_users: [%i[id role user_id user_email invite]],
+          invitation_args: %i[invite user_name id title]
+        )
       end
 
       def find_and_authorize_scenario
@@ -141,47 +118,54 @@ module Api
         end
       end
 
-      def validate_users_presence
-        return true if permitted_params[:scenario_users].present?
+      def validate_scenario_user_params(attributes)
+        if permitted_params[:scenario_users].blank?
+          render json: { error: 'No users given to perform action on.' }, status: :unprocessable_entity
 
-        render json: { error: 'No users given to perform action on.' }, status: :unprocessable_entity
+          return false
+        end
 
-        return false
-      end
+        # Check if all scenario_users have the requested attributes
+        permitted_params[:scenario_users].each do |user_params|
+          attributes.each do |attribute|
+            pp attribute
 
-      def users_return_values
-        user_ids = permitted_params[:scenario_users].pluck(:id) if permitted_params[:scenario_users].present?
+            # If this is an array of attributes, at least one of the entries should be present
+            valid = \
+              if attribute.is_a? Array
+                (attribute - user_params.keys).length < attribute.length
+              else
+                user_params.keys.include?(attribute)
+              end
 
-        scenario_users = @scenario.scenario_users
-        scenario_users = scenario_users.where(user_id: user_ids) if user_ids.present?
+            unless valid
+              missing_attr = attribute.is_a?(Array) ? attribute.join(' or ') : attribute
 
-        scenario_users.map do |u|
-          { id: u.user_id, email: u.user_email, role: User::ROLES[u.role_id] }
+              render json: user_params.merge({ error: "Missing attribute(s) for scenario_user: #{missing_attr}" }),
+                status: :unprocessable_entity
+
+              return false
+            end
+          end
         end
       end
 
-      def new_scenario_user_from(user_params)
-        user = User.find_by(email: user_params.try(:[], :email))
-
-        ScenarioUser.new(
-          scenario: @scenario,
-          role_id: User::ROLES.key(user_params.try(:[], :role).try(:to_sym)),
-          user_id: user.present? ? user.id : user_params.try(:[], :id),
-          user_email: user.present? ? nil : user_params.try(:[], :email)
-        )
-      end
-
+      # Find an existing ScenarioUser record by given user_params
       def find_scenario_user_by_params(user_params)
-        if user_params.try(:[], :id).present?
-          scenario_user = @scenario.scenario_users.find_by(user_id: user_params.try(:[], :id))
-        elsif user_params.try(:[], :email).present?
-          scenario_user = @scenario.scenario_users.find_by(user_email: user_params.try(:[], :email))
+        scenario_user = nil
 
-          # It may have happened the user was registered through e-mail but was saved by
+        if user_params[:id]&.present?
+          scenario_user = @scenario.scenario_users.find(user_params[:id])
+        elsif user_params[:user_id]&.present?
+          scenario_user = @scenario.scenario_users.find_by(user_id: user_params[:user_id])
+        elsif user_params[:user_email]&.present?
+          scenario_user = @scenario.scenario_users.find_by(user_email: user_params[:user_email])
+
+          # It may have happened the user was registered through email but was saved by
           # its ID. Search all users for this email, then search for the found ID
           # through the users for this scenario just to be sure.
           if scenario_user.blank?
-            user = User.find_by(email: user_params.try(:[], :email))
+            user = User.find_by(email: user_params[:user_email])
 
             if user.present?
               scenario_user = @scenario.scenario_users.find_by(user_id: user.id)
@@ -192,8 +176,38 @@ module Api
         scenario_user
       end
 
+      # Format the return values for the users in the given scenario
+      def users_return_values
+        if permitted_params[:scenario_users].present?
+          user_ids = permitted_params[:scenario_users].pluck(:user_id)
+        end
+
+        scenario_users = @scenario.scenario_users
+        scenario_users = scenario_users.where(user_id: user_ids) if user_ids.present?
+
+        scenario_users.map do |su|
+          { id: su.id, user_id: su.user_id, user_email: su.user_email, role: User::ROLES[su.role_id] }
+        end
+      end
+
+      # Create a new ScenarioUser record from given user_params
+      def new_scenario_user_from(user_params)
+        if user_params[:user_id].present?
+          user = User.find(user_params[:user_id])
+        elsif user_params[:user_email].present?
+          user = User.find_by(email: user_params[:email])
+        end
+
+        ScenarioUser.new(
+          scenario: @scenario,
+          role_id: User::ROLES.key(user_params.try(:[], :role).try(:to_sym)),
+          user_id: user&.id,
+          user_email: user&.email || user_params.try(:[], :user_email)
+        )
+      end
+
       # Make sure a user knows it was added to a scenario by sending an email notifying them.
-      def send_invitation_mail(scenario_user)
+      def send_invitation_mail_for(scenario_user)
         if scenario_user.user_id.present?
           user_type = 'existing'
           email = scenario_user.user.email
